Do not error when the master brand is not set.
This allows v2 to be compatible with v3.1, enabling the use of a semver range ">=2.2.0 <4".
Enabling this range means we do not need to release a major version of every component which relies on o-brand.
The assumption is no new components will use v2, therefore the error is safe to remove.